# Port from futures-util to futures-lite