--- a/src/base/read/io/entry.rs
+++ b/src/base/read/io/entry.rs
@@ -9,7 +9,7 @@ use crate::spec::Compression;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
-use futures_util::io::{AsyncRead, AsyncReadExt, BufReader, Take};
+use futures_lite::io::{AsyncRead, AsyncReadExt, BufReader, Take};
 use pin_project::pin_project;
 
 /// A type which encodes that [`ZipEntryReader`] has associated entry data.