--- a/src/base/read/io/owned.rs
+++ b/src/base/read/io/owned.rs
@@ -4,7 +4,7 @@
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
-use futures_util::io::{AsyncBufRead, AsyncRead, BufReader};
+use futures_lite::io::{AsyncBufRead, AsyncRead, BufReader};
 use pin_project::pin_project;
 
 /// A wrapping reader which holds an owned R or a mutable borrow to R.