--- a/src/base/write/entry_whole.rs
+++ b/src/base/write/entry_whole.rs
@@ -15,7 +15,7 @@ use crate::spec::{
 };
 use crate::StringEncoding;
 #[cfg(any(feature = "deflate", feature = "bzip2", feature = "zstd", feature = "lzma", feature = "xz"))]
-use futures_util::io::Cursor;
+use futures_lite::io::Cursor;
 
 use crate::spec::consts::{NON_ZIP64_MAX_NUM_FILES, NON_ZIP64_MAX_SIZE};
 #[cfg(any(feature = "deflate", feature = "bzip2", feature = "zstd", feature = "lzma", feature = "xz"))]