futures-lite is a subset of futures-util that is smaller and compiles much faster.