@Majored 2023-11-17
`futures-lite` advertises itself as much faster to compile than the full `futures` crate, but after a quick test locally, it doesn't see the same performance benefits over `futures-util`.

I'd rather stick to `futures-util` given it's actually part of the full crate.