@notgull 2023-11-18
Current `master`:

```
$ cargo clean && time cargo build -j 1 -q
cargo build -j 1 -q  10.84s user 1.81s system 95% cpu 13.251 total
```

This branch:

```
$ cargo clean && time cargo build -j 1 -q
cargo build -j 1 -q  8.76s user 1.35s system 96% cpu 10.508 total
```

After an adjustment to remove some dependencies:

```
$ cargo clean && time cargo build -j 1 -q
cargo build -j 1 -q  8.59s user 1.31s system 99% cpu 9.931 tota
```

I don't think that a 20% reduction in build times is anything to sneeze at here. Although you mention "performance", is there a benchmark that you're using here

> I'd rather stick to `futures-util` given it's actually part of the full crate.

Could you elaborate here?