@Majored 2023-11-25
> Although you mention "performance", is there a benchmark that you're using here

Ah, bad wording on my part - I was still referring to the compile speeds.

> Could you elaborate here?

I essentially just don't want to be chasing minor improvements in compile speeds by constantly switching out different dependencies and ones that may be more obscure/less maintained than others. Whilst I appreciate that `futures-lite` is part of the smol ecosystem, `futures-util` is more than suitable currently I feel - we're only pulling in what we need.