@notgull 2023-11-25
> I essentially just don't want to be chasing minor improvements

I would argue 20% is not just a minor improvement.

> ones that may be more obscure/less maintained than others

The maintainer of `futures-util` is also a maintainer of `futures-lite`. So maintainership shouldn't be an issue.

> we're only pulling in what we need.

By pulling in `futures-util`, you are also pulling in:

- An asynchronous channel implementation, alongside implementations of several asynchronous locks
- A poorly optimized `async` executor (needed for implementing concurrent combinators).
- A set of macros for joining and racing, `syn`, and therefore the entire proc macro ecosystem.
   - Note that this crate already uses `syn` via the `pin-project` crate. I was planning on porting it to `pin-project-lite` in another PR.
- A handful of combinators that are active footguns in the `async` ecosystem, but can't be removed due to causing a breaking change.
- A 5MB PNG image of Guy Fieri.

So I think that the subset of `futures-util` this crate uses is serviced better by `futures-lite`.